+++ docs/changelog.md
@@ -6,6 +6,9 @@
 
 [Скопируйте обновленный код](https://script.google.com/d/1DnC4H7yjqPV2unMZ_nmB-1bDSJT9wQUJ7Wq-ijF4Nc7Fl3qnbT0FkPSr/edit?usp=sharing).
 
+## Версия 1.5.3
+- Новая функция [checkFavoriteTracks](/func?id=checkfavoritetracks).
+
 ## Версия 1.5.2
 - Новые функции [getSavedAlbums](/func?id=getsavedalbums), [followPlaylists](/func?id=followPlaylists), [unfollowPlaylists](/func?id=unfollowPlaylists).
 - Функция [getFollowedTracks](/func?id=getfollowedtracks) получила параметр `isFlat`.